+++ studio/commands/__init__.py
@@ -3,6 +3,7 @@
 from studio.commands.command import Command
 from studio.commands.command_manager import CommandManager
 from studio.commands.delete_piece_command import DeletePieceCommand
+from studio.commands.duplicate_piece_command import DuplicatePieceCommand
 from studio.commands.import_boards_command import ImportBoardsCommand
 from studio.commands.import_pieces_command import ImportPiecesCommand
 from studio.commands.move_piece_command import MovePieceCommand
@@ -12,6 +13,7 @@ __all__ = [
     "Command",
     "CommandManager",
     "DeletePieceCommand",
+    "DuplicatePieceCommand",
     "ImportBoardsCommand",
     "ImportPiecesCommand",
     "MovePieceCommand",